Introsound geadd aan startscene
As we know, the goal of every struggle is victory. But if the
proletariat is to achieve victory, all the workers, irrespective of
nationality, must be united. Clearly, the demolition of national
barriers and close unity between the Russian, Georgian, Armenian,
Polish, Jewish and other proletarians is a necessary condition for the
victory of the proletariat of all Russia.
- Joseph Stalin